MAEBY FINISH:  se hacen cambios a funcionalidades y end point de editar  tarea y eliminar tarea   se corrigen bugs

+++ front-crud-database-kata/src/components/CategoriaTareas.jsx
@@ -11,28 +11,48 @@ const CategoryNote = () => {
   const [id, setId] = useState("");
   const [notesList, setNotesList] = useState([]);
 
-  const removeNote = (note) => {
-    dispatch({
-      type: "delete-note",
-      payload: note,
-    });
+  const removeNote = async (note) => {
+    console.log(note.id);
+    let response = await fetch(
+      `http://localhost:8080/api/delete/note/${note.id}`,
+      {
+        method: "DELETE",
+      }
+    );
+    if (response.status == 200) {
+      console.log("Borradoooo");
+
+      dispatch({
+        type: "delete-note",
+        payload: note,
+      });
+    }
   };
 
-  const removeCategory = (categoryId) => {
-    dispatch({
-      type: "delete-category",
-      payload: {
-        id: categoryId,
-      },
-    });
+  const removeCategory = async (categoryId) => {
+    let response = await fetch(
+      `http://localhost:8080/api/delete/category/${categoryId}`,
+      {
+        method: "DELETE",
+      }
+    );
+    if (response.status == 200) {
+      dispatch({
+        type: "delete-category",
+        payload: {
+          id: categoryId,
+        },
+      });
+    }
   };
 
   useEffect(() => {
-    let listOfCategories = fetchAllCategories().then((categories) => {
+    let listOfCategories = fetchAllCategories().then((category) => {
       let action = {
         type: "get-categories",
-        payload: categories,
+        payload: category,
       };
+      console.log(category);
       dispatch(action);
     });
   }, []);
